ingestor/requirements.txt
kafka-python>=2,<3
psycopg2-binary>=2.9,<3
